change package name as part of renaming
